+++ packages/init/templates/demo/src/app/constants.ts
@@ -2,19 +2,27 @@
  * 应用内常量模块
  */
 
+// 用于存储的key
 export const storeKeys = {
   auth: 'auth',
   schemaEditor: 'schemaEditor',
 }
 
+// 用于消息通知的 key
 export const msgKeys = {
   updateAuthLoginCode: 'updateAuthLoginCode',
 }
 
+// 权限校验key
+export const limitKeys = {
+  global: {
+    // 全局权限 是根据 app/routes/global 的 nodePath 与 limits 字段 key 拼接而成，其他的权限逻辑类似
+    sysRoleIdPicker: '/_global/system/sysRoleIdPicker',
+    sysUserInfoModal: '/_global/system/sysUserInfoModal',
+  },
+}
+
+// 覆盖ovine内部的常量
 export const appConstants = {
   loginRoute: '/login',
-  actionAddrMap: {
-    'POST ovapi/user/login': '管理员登录系统',
-    'POST ovapi/user/logout': '管理员退出系统',
-  },
 }