+++ packages/init/templates/demo/src/pages/dashboard/preset.ts
@@ -15,6 +15,7 @@ const prest: PagePreset = {
   apis: {
     chart: {
       url: 'GET ovapi/stat/data',
+      cache: 500,
       limits: '$page',
     },
   },